+++ platform/src/codatplatform/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '1.3.0'
-    gen_version: str = '2.139.1'
-    user_agent: str = 'speakeasy-sdk/python 1.3.0 2.139.1 3.0.0 codat-platform'
+    sdk_version: str = '1.4.0'
+    gen_version: str = '2.142.2'
+    user_agent: str = 'speakeasy-sdk/python 1.4.0 2.142.2 3.0.0 codat-platform'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: